Changed trailAscent to INT https://app.asana.com/0/660901941582663/663621423549618

+++ sql/abq-outside.sql
@@ -18,7 +18,7 @@ CREATE TABLE profile (
 CREATE TABLE trail (
 	trailId BINARY(16) NOT NULL,
 	trailAddress VARCHAR(255),
-	trailAscent TINYINT,
+	trailAscent INT,
 	trailImage VARCHAR(255),
 	trailLat DECIMAL (12,9),
 	trailLength DECIMAL (6,3),
